requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py38h32f6830_1
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h7b6447c_1000
bzip2=1.0.8=h516909a_2
c-ares=1.16.1=h516909a_0
ca-certificates=2020.6.24=0
cachetools=4.1.1=py_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38_0
cffi=1.14.0=py38hd463f26_0
chardet=3.0.4=py38_1003
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.5.0=py_0
cpuonly=1.0=0
cryptography=2.9.2=py38h1ba5d50_0
cycler=0.10.0=py38_0
cytoolz=0.10.1=py38h7b6447c_0
dask-core=2.20.0=py_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
enum34=1.1.10=py38h32f6830_1
expat=2.2.9=he6710b0_2
ffmpeg=4.3=h167e202_0
flask=1.1.2=py_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=py38h32f6830_1
gettext=0.19.8.1=hc5be6a0_1002
git=2.23.0=pl526hacde149_0
glib=2.65.0=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-auth=1.20.0=py_0
google-auth-oauthlib=0.4.1=py_2
googledrivedownloader=0.4=py38_0
graphite2=1.3.13=he1b5a44_1001
grpcio=1.30.0=py38h2c89da0_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_100
icu=64.2=he1b5a44_1
idna=2.10=py_0
imagecodecs-lite=2019.12.3=py38h8790de6_1
imageio=2.9.0=py_0
importlib-metadata=1.7.0=py38h32f6830_0
intel-openmp=2020.1=217
itsdangerous=1.1.0=py_0
jasper=1.900.1=h07fcdf6_1006
jinja2=2.11.2=py_0
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py38hfd86e86_0
krb5=1.18.2=h173b8e3_0
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.34=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=9.0.1=default_hde54327_0
libcurl=7.71.1=h20c2e04_1
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1006
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm9=9.0.1=he513fc3_1
libopencv=4.4.0=py38_2
libpng=1.6.37=hbc83047_0
libprotobuf=3.12.3=h8b12597_2
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.14=h7b6447c_0
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he6710b0_0
markdown=3.2.2=py_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.0=1
matplotlib-base=3.3.0=py38h91b0d89_1
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_1
ninja=1.9.0=py38hfd86e86_0
nspr=4.27=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.18.5=py38ha1c710e_0
numpy-base=1.18.5=py38hde5b4d6_0
oauthlib=3.0.1=py_0
olefile=0.46=py_0
opencv=4.4.0=py38_2
openh264=2.1.1=h8b12597_0
openssl=1.1.1g=h7b6447c_0
pandas=1.0.5=py38h0573a6f_0
pathlib=1.0.1=py_1
pcre=8.44=he6710b0_0
perl=5.26.2=h14c3975_0
pillow=7.2.0=py38hb39fc2d_0
pip=20.1.1=py38_1
pixman=0.38.0=h516909a_1003
protobuf=3.12.3=py38h950e882_0
py-opencv=4.4.0=py38h23f93f0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_2
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pysocks=1.7.1=py38_0
python=3.8.5=h425cb1d_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.5.1=py3.8_cpu_0
pytorch-lightning=0.8.5=py_0
pytz=2020.1=py_0
pywavelets=1.1.1=py38h7b6447c_0
pyyaml=5.3.1=py38h7b6447c_1
qt=5.12.5=hd8c4c69_1
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
scikit-image=0.17.2=py38hcb8c335_1
scikit-learn=0.23.1=py38h423224d_0
scipy=1.5.0=py38h0b6359f_0
setuptools=49.2.0=py38_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
sk-video=1.1.10=pyh24bf2e0_4
sqlite=3.32.3=h62c20be_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2019.7.26.2=py38_0
tk=8.6.10=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.6.1=py38_cpu
tornado=6.0.4=py38h7b6447c_1
tqdm=4.48.0=pyh9f0ad1d_0
urllib3=1.25.9=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py38_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h0b5b093_0
